refactor(combobox): split combobox input from trigger
