+++ packages/components/combobox/src/Combobox.doc.mdx
@@ -33,7 +33,12 @@ import { Combobox } from '@spark-ui/combobox'
     'Combobox.Trigger': {
       of: Combobox.Trigger,
       description:
-        'The button that toggles the select. The Select.Popover will position itself by aligning over the trigger.',
+        'The area that toggles the combobox popover. The Select.Popover will position itself by aligning with its trigger.',
+    },
+    'Combobox.Input': {
+      of: Combobox.Input,
+      description:
+        'The typing area in which the user can type. The input behaviour will differ if `autoComplete` is used or not.',
     },
     'Combobox.LeadingIcon': {
       of: Combobox.LeadingIcon,
