+++ packages/components/combobox/src/Combobox.stories.tsx
@@ -19,7 +19,7 @@ export default meta
  * Minimal anatomy:
  * - Combobox
  *  - Combobox.Trigger
- *    - Combobox.Trigger
+ *    - Combobox.Input
  *  - Combobox.Popover
  *    - Combobox.Items
  *      - Combobox.Item
@@ -29,7 +29,7 @@ export default meta
  *  - Combobox.Trigger
  *    - Combobox.LeadingIcon
  *    - Combobox.SelectedItems
- *    - Combobox.Trigger
+ *    - Combobox.Input
  *    - Combobox.ClearButton
  *    - Combobox.Disclosure
  *  - Combobox.Popover
@@ -64,7 +64,9 @@ export const Default: StoryFn = _args => {
   return (
     <div className="pb-[300px]">
       <Combobox>
-        <Combobox.Trigger aria-label="Book" placeholder="Pick a book" />
+        <Combobox.Trigger>
+          <Combobox.Input aria-label="Book" placeholder="Pick a book" />
+        </Combobox.Trigger>
 
         <Combobox.Popover>
           <Combobox.Items>
@@ -88,12 +90,14 @@ export const Controlled: StoryFn = () => {
   return (
     <div className="pb-[300px]">
       <Combobox value={value} onValueChange={setValue}>
-        <Combobox.Trigger
-          aria-label="Book"
-          placeholder="Pick a book"
-          value={inputValue}
-          onValueChange={setInputValue}
-        />
+        <Combobox.Trigger>
+          <Combobox.Input
+            aria-label="Book"
+            placeholder="Pick a book"
+            value={inputValue}
+            onValueChange={setInputValue}
+          />
+        </Combobox.Trigger>
 
         <Combobox.Popover>
           <Combobox.Items>
@@ -126,7 +130,9 @@ export const ControlledOpenState: StoryFn = () => {
 
       <div className="pb-[300px]">
         <Combobox open={open} onOpenChange={setOpen}>
-          <Combobox.Trigger aria-label="Book" placeholder="Pick a book" />
+          <Combobox.Trigger>
+            <Combobox.Input aria-label="Book" placeholder="Pick a book" />
+          </Combobox.Trigger>
 
           <Combobox.Popover>
             <Combobox.Items>
@@ -148,7 +154,9 @@ export const CustomItem: StoryFn = _args => {
   return (
     <div className="pb-[300px]">
       <Combobox>
-        <Combobox.Trigger aria-label="Book" placeholder="Pick a book" />
+        <Combobox.Trigger>
+          <Combobox.Input aria-label="Book" placeholder="Pick a book" />
+        </Combobox.Trigger>
 
         <Combobox.Popover>
           <Combobox.Items>
@@ -187,7 +195,9 @@ export const Disabled: StoryFn = _args => {
   return (
     <div className="pb-[300px]">
       <Combobox disabled>
-        <Combobox.Trigger aria-label="Book" placeholder="Pick a book" />
+        <Combobox.Trigger>
+          <Combobox.Input aria-label="Book" placeholder="Pick a book" />
+        </Combobox.Trigger>
 
         <Combobox.Popover>
           <Combobox.Items>
@@ -217,7 +227,9 @@ export const FilteringAutoFilter: StoryFn = _args => {
   return (
     <div className="pb-[300px]">
       <Combobox autoFilter>
-        <Combobox.Trigger aria-label="Book" placeholder="Pick a book" />
+        <Combobox.Trigger>
+          <Combobox.Input aria-label="Book" placeholder="Pick a book" />
+        </Combobox.Trigger>
 
         <Combobox.Popover>
           <Combobox.Items>
@@ -248,12 +260,14 @@ export const FilteringManual: StoryFn = () => {
   return (
     <div className="pb-[300px]">
       <Combobox>
-        <Combobox.Trigger
-          aria-label="Book"
-          placeholder="Pick a book"
-          value={inputValue}
-          onValueChange={setInputValue}
-        />
+        <Combobox.Trigger>
+          <Combobox.Input
+            aria-label="Book"
+            placeholder="Pick a book"
+            value={inputValue}
+            onValueChange={setInputValue}
+          />
+        </Combobox.Trigger>
 
         <Combobox.Popover>
           <Combobox.Items>
@@ -278,7 +292,9 @@ export const ReadOnly: StoryFn = _args => {
   return (
     <div className="pb-[300px]">
       <Combobox readOnly>
-        <Combobox.Trigger aria-label="Book" placeholder="Pick a book" />
+        <Combobox.Trigger aria-label="Book">
+          <Combobox.Input aria-label="Book" placeholder="Pick a book" />
+        </Combobox.Trigger>
 
         <Combobox.Popover>
           <Combobox.Items>
@@ -299,7 +315,9 @@ export const DisabledItem: StoryFn = _args => {
   return (
     <div className="pb-[300px]">
       <Combobox>
-        <Combobox.Trigger aria-label="Book" placeholder="Pick a book" />
+        <Combobox.Trigger>
+          <Combobox.Input aria-label="Book" placeholder="Pick a book" />
+        </Combobox.Trigger>
 
         <Combobox.Popover>
           <Combobox.Items>
@@ -322,7 +340,9 @@ export const Grouped: StoryFn = _args => {
   return (
     <div className="pb-[300px]">
       <Combobox autoFilter>
-        <Combobox.Trigger aria-label="Book" placeholder="Pick a book" />
+        <Combobox.Trigger>
+          <Combobox.Input aria-label="Book" placeholder="Pick a book" />
+        </Combobox.Trigger>
         <Combobox.Popover>
           <Combobox.Items>
             <Combobox.Group>
@@ -351,7 +371,9 @@ export const LeadingIcon: StoryFn = _args => {
   return (
     <div className="pb-[300px]">
       <Combobox>
-        <Combobox.Trigger aria-label="Book" placeholder="Pick a book" />
+        <Combobox.Trigger>
+          <Combobox.Input aria-label="Book" placeholder="Pick a book" />
+        </Combobox.Trigger>
 
         <Combobox.Popover>
           <Combobox.Items>
@@ -374,7 +396,9 @@ export const ItemIndicator: StoryFn = _args => {
   return (
     <div className="pb-[300px]">
       <Combobox multiple defaultValue={['book-1', 'book-2']}>
-        <Combobox.Trigger aria-label="Book" placeholder="Pick a book" />
+        <Combobox.Trigger aria-label="Book">
+          <Combobox.Input aria-label="Book" placeholder="Pick a book" />
+        </Combobox.Trigger>
 
         <Combobox.Popover>
           <Combobox.Items>
@@ -419,7 +443,9 @@ export const Statuses: StoryFn = () => {
       {statuses.map(status => {
         return (
           <Combobox state={status}>
-            <Combobox.Trigger aria-label="Book" placeholder="Pick a book" />
+            <Combobox.Trigger aria-label="Book">
+              <Combobox.Input aria-label="Book" placeholder="Pick a book" />
+            </Combobox.Trigger>
 
             <Combobox.Popover>
               <Combobox.Items>
@@ -442,7 +468,9 @@ export const MultipleSelection: StoryFn = _args => {
   return (
     <div className="pb-[300px]">
       <Combobox multiple autoFilter defaultValue={['book-1', 'book-2']}>
-        <Combobox.Trigger aria-label="Book" placeholder="Pick a book" />
+        <Combobox.Trigger>
+          <Combobox.Input aria-label="Book" placeholder="Pick a book" />
+        </Combobox.Trigger>
 
         <Combobox.Popover>
           <Combobox.Items>
@@ -475,12 +503,14 @@ export const MultipleSelectionControlled: StoryFn = () => {
   return (
     <div className="flex flex-col  gap-md pb-[300px]">
       <Combobox autoFilter multiple value={selectedValues} onValueChange={setSelectedValues}>
-        <Combobox.Trigger
-          aria-label="Book"
-          placeholder="todo placeholder"
-          value={inputValue}
-          onValueChange={setInputValue}
-        />
+        <Combobox.Trigger>
+          <Combobox.Input
+            aria-label="Book"
+            placeholder="Pick a book"
+            value={inputValue}
+            onValueChange={setInputValue}
+          />
+        </Combobox.Trigger>
 
         <Combobox.Popover>
           <Combobox.Items>
@@ -505,7 +535,9 @@ export const FormFieldLabel: StoryFn = _args => {
       <FormField>
         <FormField.Label>Book</FormField.Label>
         <Combobox>
-          <Combobox.Trigger aria-label="Book" placeholder="Pick a book" />
+          <Combobox.Trigger>
+            <Combobox.Input placeholder="Pick a book" />
+          </Combobox.Trigger>
           <Combobox.Popover>
             <Combobox.Items>
               <Combobox.Item value="book-1">To Kill a Mockingbird</Combobox.Item>
@@ -530,7 +562,9 @@ export const FormFieldHiddenLabel: StoryFn = _args => {
           <VisuallyHidden>Book</VisuallyHidden>
         </FormField.Label>
         <Combobox>
-          <Combobox.Trigger aria-label="Book" placeholder="Pick a book" />
+          <Combobox.Trigger>
+            <Combobox.Input placeholder="Pick a book" />
+          </Combobox.Trigger>
           <Combobox.Popover>
             <Combobox.Items>
               <Combobox.Item value="book-1">To Kill a Mockingbird</Combobox.Item>
@@ -553,7 +587,9 @@ export const FormFieldReadOnly: StoryFn = _args => {
       <FormField readOnly>
         <FormField.Label>Book</FormField.Label>
         <Combobox>
-          <Combobox.Trigger aria-label="Book" placeholder="Pick a book" />
+          <Combobox.Trigger>
+            <Combobox.Input aria-label="Book" placeholder="Pick a book" />
+          </Combobox.Trigger>
 
           <Combobox.Popover>
             <Combobox.Items>
@@ -577,7 +613,9 @@ export const FormFieldDisabled: StoryFn = _args => {
       <FormField disabled>
         <FormField.Label>Book</FormField.Label>
         <Combobox>
-          <Combobox.Trigger aria-label="Book" placeholder="Pick a book" />
+          <Combobox.Trigger>
+            <Combobox.Input aria-label="Book" placeholder="Pick a book" />
+          </Combobox.Trigger>
           <Combobox.Popover>
             <Combobox.Items>
               <Combobox.Item value="book-1">To Kill a Mockingbird</Combobox.Item>
@@ -600,7 +638,9 @@ export const FormFieldRequired: StoryFn = _args => {
       <FormField isRequired>
         <FormField.Label>Book</FormField.Label>
         <Combobox>
-          <Combobox.Trigger aria-label="Book" placeholder="Pick a book" />
+          <Combobox.Trigger>
+            <Combobox.Input aria-label="Book" placeholder="Pick a book" />
+          </Combobox.Trigger>
           <Combobox.Popover>
             <Combobox.Items>
               <Combobox.Item value="book-1">To Kill a Mockingbird</Combobox.Item>
@@ -629,7 +669,9 @@ export const FormFieldValidation: StoryFn = () => {
             setState(value === 'default' ? undefined : (value as 'success' | 'alert' | 'error'))
           }}
         >
-          <Combobox.Trigger aria-label="Book" placeholder="Pick a book" />
+          <Combobox.Trigger aria-label="Book">
+            <Combobox.Input aria-label="Book" placeholder="Pick a book" />
+          </Combobox.Trigger>
           <Combobox.Popover>
             <Combobox.Items>
               <Combobox.Item value="default">default</Combobox.Item>
